Se migro el modelo sexo de legajos a configuraciones y todo lo relacionado a el.(fixture, admin)
