+++ configuraciones/models.py
@@ -708,3 +708,14 @@ class Vacantes(models.Model):
 
     def get_absolute_url(self):
         return reverse("vacantes_ver", kwargs={"pk": self.pk})
+
+
+class Sexo(models.Model):
+    sexo = models.CharField(max_length=10)
+
+    def __str__(self):
+        return str(self.sexo)
+
+    class Meta:
+        verbose_name = "Sexo"
+        verbose_name_plural = "Sexos"
